requirements.txt
# Framework web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Validación de datos
pydantic>=2.5.0

# Cliente HTTP
requests>=2.31.0
urllib3>=2.1.0

# Smartsheet SDK
smartsheet-python-sdk>=3.0.0

# Pandas para manipulación de datos
pandas>=2.1.0

# Scheduler
APScheduler>=3.10.0

# Variables de entorno
python-dotenv>=1.0.0

# Logging y utilidades
python-json-logger>=2.0.0
